+++ src/main/java/app/AppBuilder.java
@@ -28,15 +28,15 @@ import interface_adapter.signup.club_signup.ClubSignupViewModel;
 import interface_adapter.signup.student_signup.StudentSignupController;
 import interface_adapter.signup.student_signup.StudentSignupPresenter;
 import interface_adapter.signup.student_signup.StudentSignupViewModel;
-import interface_adapter.student_home.StudentHomeController;
-import interface_adapter.student_home.StudentHomePresenter;
-import interface_adapter.student_home.StudentHomeViewModel;
-import interface_adapter.student_home.dislike.StudentDislikeController;
-import interface_adapter.student_home.like.StudentLikeController;
-import interface_adapter.student_home.show_clubs.StudentShowClubsController;
-import interface_adapter.student_home.show_clubs.StudentShowClubsPresenter;
-import interface_adapter.student_home.show_posts.StudentShowPostsController;
-import interface_adapter.student_home.show_posts.StudentShowPostsPresenter;
+import interface_adapter.student_logged_in.student_home.StudentHomeController;
+import interface_adapter.student_logged_in.student_home.StudentHomePresenter;
+import interface_adapter.student_logged_in.student_home.StudentHomeViewModel;
+import interface_adapter.student_logged_in.student_home.dislike.StudentDislikeController;
+import interface_adapter.student_logged_in.student_home.like.StudentLikeController;
+import interface_adapter.student_logged_in.student_home.show_clubs.StudentShowClubsController;
+import interface_adapter.student_logged_in.student_home.show_clubs.StudentShowClubsPresenter;
+import interface_adapter.student_logged_in.student_home.show_posts.StudentShowPostsController;
+import interface_adapter.student_logged_in.student_home.show_posts.StudentShowPostsPresenter;
 import interface_adapter.student_profile.StudentProfileController;
 import interface_adapter.student_profile.StudentProfilePresenter;
 import interface_adapter.student_profile.StudentProfileViewModel;
