--- src/main/java/interface_adapter/student_home/search/StudentSearchController.java
@@ -1,4 +1,0 @@
-package interface_adapter.student_home.search;
-
-public class StudentSearchController {
-}
