+++ src/main/java/interface_adapter/student_logged_in/student_home/StudentHomeController.java
@@ -1,4 +1,4 @@
-package interface_adapter.student_home;
+package interface_adapter.student_logged_in.student_home;
 
 import use_case.student_homepage.StudentHomeInputBoundary;
