+++ src/main/java/interface_adapter/student_logged_in/student_home/search/StudentSearchController.java
@@ -1,0 +1,4 @@
+package interface_adapter.student_logged_in.student_home.search;
+
+public class StudentSearchController {
+}
