+++ src/main/java/view/StudentHomeView.java
@@ -14,13 +14,13 @@ import javax.swing.JTextField;
 import javax.swing.event.DocumentEvent;
 import javax.swing.event.DocumentListener;
 
-import interface_adapter.student_home.StudentHomeController;
-import interface_adapter.student_home.StudentHomeState;
-import interface_adapter.student_home.StudentHomeViewModel;
-import interface_adapter.student_home.dislike.StudentDislikeController;
-import interface_adapter.student_home.like.StudentLikeController;
-import interface_adapter.student_home.show_clubs.StudentShowClubsController;
-import interface_adapter.student_home.show_posts.StudentShowPostsController;
+import interface_adapter.student_logged_in.student_home.StudentHomeController;
+import interface_adapter.student_logged_in.student_home.StudentHomeState;
+import interface_adapter.student_logged_in.student_home.StudentHomeViewModel;
+import interface_adapter.student_logged_in.student_home.dislike.StudentDislikeController;
+import interface_adapter.student_logged_in.student_home.like.StudentLikeController;
+import interface_adapter.student_logged_in.student_home.show_clubs.StudentShowClubsController;
+import interface_adapter.student_logged_in.student_home.show_posts.StudentShowPostsController;
 
 /**
  * This is the view for the homepage that the student will be greated with upon logging in. This will show all the clubs
